Render placeholder tileset image for Platforms.
Our map is generated expecting Flixel#Tilemap's auto-layout mode, which
changes 1's based on their neighbors. We'll need to make up for Phaser's
lack of this logic (seemingly deferring to tilemap editors), but it's
also a good opportunity to make something more flexible than Flixel's
logic.

So for now, switch to the Flixel-esque placeholder.